Create an ExternalPreimage indicator on Invoices
This modifies the Invoice schema to include an indicator of
whether the preimage for a given invoice is stored locally or
if it is stored externally.

For externally stored preimages, the Payment Hash of the Invoice
is persisted with it, since it needs to be accessed by other users
and it can no longer be generated from the PaymentPreimage.