Add gRPC client for retrieving external preimages
This change adds a gRPC protobuf file and client to lnd for retrieving
external preimages (see #3).

It also adds a thin wrapper around the gRPC client in order to lazily
connect to the server and reuse connections where possible. It also
wraps the streaming API of the request feel more like a request/response
pattern.

It initializes this client in the server initialization, and provides it
to each created Link as part of the ChannelLinkConfig, in anticipation of
its use in the exitHop.