Request preimages for External Service when ExternalPreimage invoices are found
This change updates the behavior of the HTLCSwitch when it is the destination
for a given HTLC.

When the invoice associated with the HTLC is marked with ExternalPreimage, the
switch uses the extpreimageClient to query the configured External Preimage
service for the invoice. Only if the invoice fails to be found in this external
service does the settlement fail.

To ensure that we are only settling with valid preimages, the extpreimage client
now performs a check that the returned preimage matches the provided hash.

In order to facillitate the Symbol parameter on the external preimage service,
the extpreimage client was made a bit thicker to make use of it easier and
giving the lnd server the ability to configure the symbol at startup.

This change introduces a few new testing helpers which are near replicas of the
helpers already in LND. This is to contain our changes as much as possible to
make this fork easier to maintain.

This change also makes the extpreimage testing more robust by creating
expectations for requests structured in a particular way.