Add local Preimages to invoices with External Preimages
This change exposes a method on the channel db and invoice registry
to add and save a preimge to an invoice that was previously marked
as having an external preimage.

Calling this method once we've retrieved the invoice from the external
service allows the invoice to have all of its properties once it is
settled and simplifies accepting duplicate payments.