Use local preimages if available, and save external preimages before settling
This change makes use of the AddInvoicePreimage methods added to the ChannelDB
and InvoiceRegistry to persist the externally retrieved preimage before
settling the invoice that requires it.

It also changes the handling of preimages so that if an external preimage has
been previously retrieved, it can be re-used for duplicate payments without
making another request to the external service.