Add External Preimage flag to invoices when creating in RPC Server
This change fixes a bug where an external preimage flag being set
in the RPC request was not being set on the channeldb Invoice.

This also adds an integration test for adding external preimage
invoices to at least ensure that creating invoices works as
expected.

This still leaves actually executing external preimage invoices
to a future integration test.